Fix All Clippy Errors (97 → 0)
Major Fixes:
- large_enum_variant: Boxed large variants in Matter and CondensationResult enums
- result_large_err: Boxed large error variant in DensityTransitionError
- too_many_arguments: Added #[allow] or created config structs for 20+ functions
- inherent_to_string: Implemented Display trait for NucleotideSequence and Codon
- type_complexity: Created type aliases for 8 complex types

Code Quality Improvements:
- assertions_on_constants: Moved 9 assertions to const blocks
- nonminimal_bool: Simplified boolean expressions
- redundant_closure_call: Removed unnecessary closures
- manual_find: Replaced manual loops with iterator find
- manual_slice_fill: Used fill() method
- manual_memcpy: Used copy_from_slice()
- len_zero: Used is_empty() instead of len() == 0
- unused_enumerate_index: Removed unnecessary enumerate calls
- wrong_self_convention: Fixed to_* methods for Copy types
- missing_safety_doc: Added # Safety sections to unsafe functions
- range_plus_one: Converted OR patterns to range syntax
- wildcard_enum_match_arm: Fixed wildcard pattern issues

API Changes:
- LayerBuilder.add() renamed to push_layer() to avoid Add trait confusion
- TradeAgreement::new() now takes TradeAgreementParams struct
- AtomFormationEvent::new() now takes AtomManifestationConfig struct
- SubSubLogos now has builder pattern for entity creation

Dead Code:
- Added #[allow(dead_code)] to test helper structs and functions

+++ src/simulation_v3/holographic_game_loop.rs
@@ -589,7 +589,7 @@ mod tests {
             let result = game_loop.run_step(1.0);
             assert_eq!(result.step, i);
             assert_eq!(game_loop.current_step(), i);
-            assert_eq!(game_loop.current_time(), i as u64);
+            assert_eq!(game_loop.current_time(), i);
         }
     }
 }
